% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CFB_json_from_url.R
\name{get_CFB_json_from_url}
\alias{get_CFB_json_from_url}
\title{Makes call to CollegeFootballData.com API and returns the JSON from the response}
\usage{
get_CFB_json_from_url(urlPath, queryParams = NULL,
  cacheLocation = NULL, cacheTimeSec = 0)
}
\arguments{
\item{urlPath}{The path (no root) for the desired resource. Venues would pass in 'venues'.}

\item{queryParams}{A named list with the desired query parameters. list(year=2016, week=1) will result in ?year=2016&week=1 being used as the query string.}

\item{cacheLocation}{A check for a cached JSON response. This method does not currently cache on its own.}

\item{cacheTimeSec}{Time in seconds that will result in the cache being ignored. A value of 30 will result in cached values being ignored if created over 30 seconds ago. A value of 0 mean no expiration.}
}
\description{
Gets desired data as JSON
}
\examples{

}
