% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coach_seasons.R
\name{get_coach_seasons}
\alias{get_coach_seasons}
\title{Gets all available coach season data}
\usage{
get_coach_seasons(firstName = NULL, lastName = NULL, team = NULL,
  year = NULL, minYear = NULL, maxYear = NULL)
}
\arguments{
\item{firstName}{(optional) First Name of coach}

\item{lastName}{(optional) Last Name of coach}

\item{team}{(optional) Name for team. "Michigan" and 'Oregon State' are valid examples.}

\item{year}{(optional) Individual year value}

\item{minYear}{(optional) Starting year for which coaches should be returned. Will be ignored if year is provided.}

\item{maxYear}{(optional) Final year for which coaches should be returned. Will be ignored if year is provided.}
}
\description{
Gets all available coach season data
}
\examples{

}
