% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conferences.R
\name{get_conferences}
\alias{get_conferences}
\title{Gets all supported conferences (current and historical)}
\usage{
get_conferences()
}
\description{
Gets all supported conferences (current and historical)
}
\examples{

}
