% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drives.R
\name{get_drives}
\alias{get_drives}
\title{Gets a set of college football drives}
\usage{
get_drives(year, week = NULL, seasonType = "regular", team = NULL,
  offense = NULL, defense = NULL, conference = NULL,
  offenseConference = NULL, defenseConference = NULL)
}
\arguments{
\item{year}{(required) Year for desired games}

\item{week}{(optional) Week number}

\item{seasonType}{(optional) [regular/postseason] Season Type for desired games}

\item{team}{(optional) Name for team. "Michigan" and 'Oregon State' are valid examples.}

\item{offense}{(optional) Name for team on offense. "Michigan" and 'Oregon State' are valid examples.}

\item{defense}{(optional) Name for team on defense. "Michigan" and 'Oregon State' are valid examples.}

\item{offenseConference}{(optional) Abbreviation for conference of team on defense. "B1G", "PAC", and "CUSA" are valid examples.}

\item{defenseConference}{(optional) Abbreviation for conference of team on defense. "B1G", "PAC", and "CUSA" are valid examples.}
}
\description{
Get all of the drives that meet the provided criteria
}
\examples{

}
