% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_games.R
\name{get_games}
\alias{get_games}
\title{Gets a set of college football games}
\usage{
get_games(year, week = NULL, seasonType = NULL, team = NULL,
  home = NULL, away = NULL, conference = NULL)
}
\arguments{
\item{year}{(required) Year for desired games}

\item{week}{(optional) Week number}

\item{seasonType}{(optional) [regular/postseason] Season Type for desired games}

\item{home}{(optional) Name for home team. "Michigan" and 'Oregon State' are valid examples.}

\item{away}{(optional) Name for away team. "Michigan" and 'Oregon State' are valid examples.}

\item{conference}{(optional) Abbreviation for conference. "B1G", "PAC", and "CUSA" are valid examples.}
}
\description{
Get all of the games that meet the provided criteria
}
\examples{

}
