% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_games.R
\name{get_player_games}
\alias{get_player_games}
\title{Gets a set of college football games}
\usage{
get_player_games(gameId = NULL, year = NULL, seasonType = NULL,
  week = NULL, team = NULL, conference = NULL)
}
\arguments{
\item{gameId}{(optional) Single desired game ID. Required if year is not provided.}

\item{year}{(optional) Year for desired games Required if gameId is not provided.}

\item{seasonType}{(optional) [regular/postseason] Defaults to regular if not provided but week is provided.}

\item{week}{(optional) Week number. Required if year is provided but team and conference are not.}

\item{team}{(optional) Name for desired team. Results will include game stats for both the provided team and their opponents. Required if year is provided but week and conference are not.}

\item{conference}{(optional) Name for desired team. Results will include game stats for both conference teams and their opponents. Required if year is provided but week and team are not.}
}
\description{
Get all of the games that meet the provided criteria
}
\examples{

}
