% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plays.R
\name{get_plays}
\alias{get_plays}
\title{Gets a set of college football plays}
\usage{
get_plays(year, week, seasonType = "regular", team = NULL,
  offense = NULL, defense = NULL, conference = NULL,
  offenseConference = NULL, defenseConference = NULL,
  playTypeId = NULL)
}
\arguments{
\item{year}{(required) Year for desired games}

\item{week}{(required) Week number}

\item{seasonType}{(optional) [regular/postseason] Season Type for desired games}

\item{team}{(optional) Name for team. "Michigan" and 'Oregon State' are valid examples.}

\item{offense}{(optional) Name for team on offense. "Michigan" and 'Oregon State' are valid examples.}

\item{defense}{(optional) Name for team on defense. "Michigan" and 'Oregon State' are valid examples.}

\item{offenseConference}{(optional) Abbreviation for conference of team on defense. "B1G", "PAC", and "CUSA" are valid examples.}

\item{defenseConference}{(optional) Abbreviation for conference of team on defense. "B1G", "PAC", and "CUSA" are valid examples.}

\item{playTypeId}{(optional) playTypeId for desired type of play to return}
}
\description{
Get all of the plays from games that meet the provided criteria
}
\examples{

}
