% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rankings.R
\name{get_rankings}
\alias{get_rankings}
\title{Gets all available rankings for the requested year}
\usage{
get_rankings(year)
}
\arguments{
\item{year}{(required) Year for desired rankings}
}
\description{
Gets all available rankings for the requested year
}
\examples{

}
