% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roster.R
\name{get_roster}
\alias{get_roster}
\title{Gets the current roster for a requested team}
\usage{
get_roster(team)
}
\arguments{
\item{team}{(required) Name for desired team. "Michigan" and 'Oregon State' are valid examples.}
}
\description{
Gets the current roster for a requested team
}
\examples{

}
