% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_matchups.R
\name{get_team_matchups}
\alias{get_team_matchups}
\title{Gets games between two teams}
\usage{
get_team_matchups(team1, team2, minYear = NULL, maxYear = NULL)
}
\arguments{
\item{team1}{(required) Name for team. "Michigan" and 'Oregon State' are valid examples.}

\item{team2}{(required) Name for team. "Michigan" and 'Oregon State' are valid examples.}

\item{minYear}{(optional) Starting year for which matchups should be returned.}

\item{maxYear}{(optional) Final year for which matchups should be returned}
}
\description{
Get all of the teams that meet the provided criteria
}
\examples{

}
