% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_talent.R
\name{get_team_talent}
\alias{get_team_talent}
\title{Gets the calculated team talent for all teams in a desired year.}
\usage{
get_team_talent(year)
}
\arguments{
\item{year}{(required)}
}
\description{
Gets the calculated team talent for all teams in a desired year.
}
\examples{

}
