% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams.R
\name{get_teams}
\alias{get_teams}
\title{Gets a set of college football teams}
\usage{
get_teams(conference = NULL)
}
\arguments{
\item{conference}{(optional) Abbreviation for conference. "B1G", "PAC", and "CUSA" are valid examples.}
}
\description{
Get all of the teams that meet the provided criteria
}
\examples{

}
